LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.measure.emHeights-low-ascent-expected.txt

2d.text.measure.emHeights-low-ascent

Testing emHeights with reduced ascent metric


FAIL Testing emHeights with reduced ascent metric assert_equals: ctx.measureText('A').emHeightAscent === 20 (got 69[number], expected 20[number]) expected 20 but got 69
